+++ managers/NPC_Animation_Manager.py
